Add a test; ensure logging and building work
This commit adds a test that checks the logs of the application. To do
so, it also sets up logging rules for test and production.

Additional configuration is needed to run this application as a native
executable. This is added in the resources along with documentation of
how it has been generated.